Load Client List on demand
