+++ service-worker.js
@@ -252,6 +252,7 @@ const APP_SHELL = [
   '/js/changelog.js',
   '/js/client-list-runtime.js',
   '/js/client-list.js',
+  '/js/client-list-impl.js',
   '/js/nav-runtime.js',
   '/js/nav.js',
   '/js/views-router-runtime.js',
