+++ tests/test-profile-share.js
@@ -134,8 +134,8 @@ assert('Share profile has a top-level header action',
   read('js/shell-actions.js').includes('shellProfileShareDeps.openProfileShareModal()') &&
   read('js/app-shell-hooks.js').includes('configureShellProfileShareDeps({ openProfileShareModal });'));
 assert('Mobile share path moves into client list menu',
-  read('js/client-list.js').includes("label: 'Share Profile', action: 'share-profile'") &&
-  read('js/client-list.js').includes('clientListRuntime.openProfileShareModal(id)') &&
+  read('js/client-list-impl.js').includes("label: 'Share Profile', action: 'share-profile'") &&
+  read('js/client-list-impl.js').includes('clientListRuntime.openProfileShareModal(id)') &&
   read('js/app-shell-hooks.js').includes('configureClientListRuntime') &&
   read('js/app-shell-hooks.js').includes('openProfileShareModal') &&
   /@media \(max-width: 768px\), \(pointer: coarse\)[\s\S]*\.header-share-btn\s*\{[\s\S]*display:\s*none/.test(redesignCss));
