+++ tests/test-quality-guardrails.js
@@ -155,6 +155,7 @@ const highValueCheckJsModules = [
   'js/api-venice.js',
   'js/client-list-runtime.js',
   'js/client-list.js',
+  'js/client-list-impl.js',
   'js/dashboard-lab-widget-renderers.js',
   'js/dashboard-widget-renderers.js',
   'js/dna.js',
