+++ tests/test-wearables-manual.js
@@ -96,7 +96,7 @@ assert('wearables manual form UI delegates tag chip actions',
     manualFormUiSrc.includes('installWearablesManualFormDelegates') &&
     !inlineHandlerRe.test(manualFormUiSrc));
 
-const clSrc = await fetch('js/client-list.js').then(r => r.text());
+const clSrc = await fetch('js/client-list-impl.js').then(r => r.text());
 assert('Edit Client modal no longer renders the cl-bio-weight container',
   !clSrc.includes('id="cl-bio-weight"'));
 assert('Edit Client modal no longer renders the cl-bio-bp container',
